Review plugin code for MOODLE_40
