+++ classes/output/courseformat/content.php
@@ -1,0 +1,45 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Contains the default content output class.
+ *
+ * @package   format_softcourse
+ * @copyright Sylvain | Pimenko 2021 <contact@pimneko.com>
+ * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace format_softcourse\output\courseformat;
+
+use core_courseformat\output\local\content as content_base;
+
+/**
+ * Base class to render a course content.
+ *
+ * @package   format_softcourse
+ * @copyright Sylvain | Pimenko 2021 <contact@pimneko.com>
+ * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class content extends content_base {
+
+    /**
+     * @var bool Topic format has add section after each topic.
+     *
+     * The responsible for the buttons is core_courseformat\output\local\content\section.
+     */
+    protected $hasaddsection = true;
+
+}
