+++ format.php
@@ -42,7 +42,12 @@ course_create_sections_if_missing($course, 0);
 
 $renderer = $PAGE->get_renderer('format_softcourse');
 
-$renderer->print_multiple_section_page($course, null, null, null, null);
+$format = course_get_format($course);
+$outputclass = $format->get_output_classname('content');
+$widget = new $outputclass($format);
+
+echo $renderer->render($widget);
+
+$PAGE->requires->js_call_amd('format_softcourse/actions', 'init');
 
 // Include course format js module.
-$PAGE->requires->js_call_amd('format_softcourse/actions', 'init');
