+++ lang/en/format_softcourse.php
@@ -18,7 +18,7 @@
  * Strings for component 'format_softcourse', language 'en', branch 'MOODLE_35_STABLE'
  *
  * @package   format_softcourse
- * @copyright 2018 Pimenko <contact@pimneko.com>
+ * @copyright 2021 Pimenko <contact@pimneko.com>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
