+++ package.json
@@ -28,9 +28,9 @@
     "eslint-plugin-babel": "5.3.0",
     "eslint-plugin-promise": "4.2.1",
     "fb-watchman": "2.0.1",
-    "gherkin-lint": "^4.2.1",
+    "gherkin-lint": "^4.2.2",
     "glob": "7.1.6",
-    "grunt": "^1.4.0",
+    "grunt": "^1.4.1",
     "grunt-babel": "8.0.0",
     "grunt-contrib-uglify": "4.0.1",
     "grunt-contrib-uglify-es": "^3.3.0",
@@ -38,7 +38,7 @@
     "grunt-eslint": "22.0.0",
     "grunt-sass": "3.1.0",
     "grunt-stylelint": "0.15.0",
-    "jshint": "^2.11.0",
+    "jshint": "^2.13.1",
     "node-sass": "^4.14.0",
     "semver": "7.3.2",
     "shifter": "https://github.com/andrewnicols/shifter/archive/v1.2.0_moodle-0.tar.gz",
