+++ templates/section.mustache
@@ -1,0 +1,35 @@
+<li role="tab" id="softcourse-section-{{ id }}" class="softcourse-section card position-relative">
+    <a href="{{#first_cm_url}}{{ first_cm_url }}{{/first_cm_url}}" class="stretched-link">
+        <div class="softcourse-section-img {{#urlimg}} not-empty with_gradient{{/urlimg}}"
+            {{#urlimg}} style="background-image: url({{ urlimg }})" {{/urlimg}}>
+            <span><i class="fa fa-play"></i>  <span>{{ start }}</span></span>
+            {{#update_img}}
+                <i class="fa fa-edit update-img" data-toggle="tooltip" title="{{ update_img }}">
+                    <input type="file" data-sectionid="{{ id }}" data-courseid="{{ courseid }}"
+                           class="section-file">
+                </i>
+
+                <i class="fa fa-trash delete-img" data-toggle="tooltip" title="{{ delete_img }}">
+                    <input type="button" data-sectionid="{{ id }}" data-courseid="{{ courseid }}"
+                           class="section-delete-file">
+                </i>
+            {{/update_img}}
+        </div>
+        <div class="softcourse-section-body">
+            {{#name}}<span class="softcourse-section-name">{{{ name }}}</span>{{/name}}
+            {{#summary}}
+                <span class="softcourse-section-summary">{{{ summarytext }}}</span>
+            {{/summary}}
+            <div class="softcourse-section-info">
+                <i class="fa fa-leanpub" data-toggle="tooltip" title="{{{ countactivitiestooltip }}}"></i>
+                <span class="softcourse-activities-count">{{{ countactivities }}}</span>
+                {{#progression}}
+                    <div class="progress" data-toggle="tooltip" title="{{ progression }}">
+                        <div class="progress-bar" role="progressbar" style="width: {{ progression_percent }}%" aria-valuenow="{{ progression_percent }}" aria-valuemin="0" aria-valuemax="100"></div>
+                    </div>
+                    <span class="progress-percent">{{{ progression_percent }}}%</span>
+                {{/progression}}
+            </div>
+        </div>
+    </a>
+</li>
