+++ templates/sections.mustache
@@ -1,99 +1,16 @@
 <div id="softcourse-sections">
     <ul id="softcourse-tabs">
+        {{$ initialsection }}
+            {{#initialsection}}
+                {{> format_softcourse/section }}
+            {{/initialsection}}
+        {{/ initialsection}}
         {{$ sections }}
             {{# sections}}
-                <li role="tab" id="softcourse-section-{{ id }}" class="softcourse-section card position-relative">
-                    <a href="{{#first_cm_url}}{{ first_cm_url }}{{/first_cm_url}}" class="stretched-link"></a>
-                    <div class="softcourse-section-img {{#urlimg}} not-empty with_gradient{{/urlimg}}"
-                        {{#urlimg}} style="background-image: url({{ urlimg }})" {{/urlimg}}>
-                        <span><i class="fa fa-play"></i>  <span>{{ start }}</span></span>
-                        {{#update_img}}
-                            <i class="fa fa-edit update-img" data-toggle="tooltip" title="{{ update_img }}">
-                                <input type="file" data-sectionid="{{ id }}" data-courseid="{{ courseid }}"
-                                       class="section-file">
-                            </i>
-
-                            <i class="fa fa-trash delete-img" data-toggle="tooltip" title="{{ delete_img }}">
-                                <input type="button" data-sectionid="{{ id }}" data-courseid="{{ courseid }}"
-                                       class="section-delete-file">
-                            </i>
-                        {{/update_img}}
-                    </div>
-                    <div class="softcourse-section-body">
-                        <span class="softcourse-section-name">{{{ name }}}</span>
-                        <span class="softcourse-section-summary">{{{ summary }}}</span>
-                        <div class="softcourse-section-info">
-                            <i class="fa fa-leanpub" data-toggle="tooltip" title="{{{ countactivitiestooltip }}}"></i>
-                            <span class="softcourse-activities-count">{{{ countactivities }}}</span>
-                            {{#progression}}
-                                <div class="progress" data-toggle="tooltip" title="{{ progression }}">
-                                    <div class="progress-bar" role="progressbar" style="width: {{ progression_percent }}%" aria-valuenow="{{ progression_percent }}" aria-valuemin="0" aria-valuemax="100"></div>
-                                </div>
-                                <span class="progress-percent">{{{ progression_percent }}}%</span>
-                            {{/progression}}
-                        </div>
-                    </div>
-                </li>
+                {{^skip}}
+                    {{> format_softcourse/section }}
+                {{/skip}}
             {{/sections}}
         {{/ sections}}
     </ul>
 </div>
-
-{{#js}}
-    // Handle text truncate for summary.
-    // This value will be use for backup text.
-    var fulltext = [];
-    texttruncate('softcourse-section-summary');
-    function texttruncate(id) {
-        var selector = document.getElementsByClassName(id);
-        if (selector != null && selector.length > 0) {
-            var truncate = false;
-
-            for (var i = 0; i < selector.length; i++) {
-                var el = selector[i];
-                if (el != null && el.innerHTML.length > 0) {
-                    // Backup text.
-                    var wordArray = '';
-                    if (fulltext[i]) {
-                        if(el.innerHTML != fulltext[i].join(' ')) {
-                            el.innerHTML = fulltext[i].join(' ');
-                        }
-                        wordArray = fulltext[i].slice();
-                    } else {
-                        wordArray = el.innerHTML.split(' ');
-                        // Use slic() to prevent copy reference.
-                        fulltext[i] = wordArray.slice();
-                    }
-                    while(el.scrollHeight > el.offsetHeight) {
-                        wordArray.pop();
-                        el.innerHTML = wordArray.join(' ');
-                        truncate = true;
-                    }
-                    if(truncate) {
-                        document.querySelector('.softcourse-section-summary div :last-child').append('...');
-                    }
-                }
-            }
-        }
-    }
-
-    var rtime;
-    var timeout = false;
-    var delta = 200;
-    $(window).resize(function() {
-    rtime = new Date();
-    if (timeout === false) {
-    timeout = true;
-    setTimeout(resizeend, delta);
-    }
-    });
-
-    function resizeend() {
-    if (new Date() - rtime < delta) {
-    setTimeout(resizeend, delta);
-    } else {
-    timeout = false;
-    texttruncate('softcourse-section-summary');
-    }
-    }
-{{/js}}
