+++ version.php
@@ -18,13 +18,13 @@
  * Version details
  *
  * @package    format_softcourse
- * @copyright  2018 Pimenko <contact@pimneko.com>
+ * @copyright  2021 Pimenko <contact@pimneko.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->version = 2021102115;        // The current plugin version (Date: YYYYMMDDXX).
+$plugin->version = 2021102117;        // The current plugin version (Date: YYYYMMDDXX).
 $plugin->requires = 2018050800;        // Requires this Moodle version.
 $plugin->component = 'format_softcourse';    // Full name of the plugin (used for diagnostics).
 $plugin->release = '4.00rc1.0';
